Stick everything CUDA related into a header for now
